Scala 2.13 changes for neo4j-cypher module ()
* Scala 2.13 changes for neo4j-cypher

* Mostly collections related changes
* Fix deprecations
* Some addition of () to functions
* Not sure what has happened to ScalaCheck configuration in PatternGen the later Scalatest but the new approach seems much worse
* The TranslateExceptionMacros needed a fix to prevent it creating infinite recursion

* Redo PatternGen.scala as per review suggestions

+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/planmatching/PlanMatcher.scala
@@ -431,7 +431,7 @@ case class ExactPlan(name: Option[PlanNameMatcher] = None,
     val variablesDesc = variables.fold(Set.empty[PrettyString])(_.expected.map(asPrettyString.raw))
     val lhsDesc = lhs.map(_.toPlanDescription)
     val rhsDesc = rhs.map(_.toPlanDescription)
-    val estRowArg = estimatedRows.map(m => EstimatedRows(m.expectedValue, None)).toSeq
+    val estRowArg = estimatedRows.map(m => EstimatedRows(m.expectedValue.toDouble, None)).toSeq
     val rowArg = rows.map(m => Rows(m.expectedValue)).toSeq
     val timeArg = time.map(m => Time(m.expectedValue)).toSeq
     val memoryArg = memory.map(m => Memory(m.expectedValue)).toSeq
